+++ crates/header-translator/src/id.rs
@@ -353,6 +353,13 @@ impl ItemIdentifier {
         }
     }
 
+    pub fn iokit_typedefs(name: impl Into<String>) -> Self {
+        Self {
+            name: name.into(),
+            location: Location::new("__builtin__"),
+        }
+    }
+
     pub fn copyhelper(mutable: bool) -> Self {
         let name = if mutable {
             "MutableCopyingHelper"