+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `ImageIO` / `objc2-image-io`.
   - `Intents` / `objc2-intents`.
   - `IntentsUI` / `objc2-intents-ui`.
+  - `IOBluetooth` / `objc2-io-bluetooth`.
   - `JavaScriptCore` / `objc2-java-script-core`.
   - `MediaAccessibility` / `objc2-media-accessibility`.
   - `MediaToolbox` / `objc2-media-toolbox`.